refactor: 미니 게임에 대한 매니저 및 소스 수정
1. 연필 깎는 미니 게임에 대한 매니저 수정 및 소스 수정
2. 흑연심 분리하는 미니 게임에 대한 소스 추가 및 수정

+++ Assets/Script/Tool/BaseTool.cs
@@ -4,7 +4,7 @@ public abstract class BaseTool : MonoBehaviour
 {
     [Header("공통 도구 스탯")]
     [SerializeField] string toolName;
-    [SerializeField] public float shavePower; // 깎는 힘
+    [SerializeField] public int shavePower; // 깎는 힘
 
     public abstract void TryUseTool(SharpeningPencil targetPencil);
 }
